@@ -13,19 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.15",
+    "@angular/animations": "8.2.14",
     "@angular/cdk": "^7.1.0",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/core": "7.2.15",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
     "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "7.2.15",
-    "@angular/http": "7.2.15",
+    "@angular/forms": "8.2.14",
     "@angular/material": "^7.1.0",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
-    "@angular/service-worker": "7.2.15",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@angular/service-worker": "8.2.14",
     "@ngx-translate/core": "^10.0.2",
     "@ngx-translate/http-loader": "^3.0.1",
     "@stomp/ng2-stompjs": "^7.2.0",
@@ -36,22 +35,22 @@
     "material-design-icons": "^3.0.1",
     "ng2-charts": "^1.6.0",
     "ngx-markdown": "^8.2.2",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.4",
     "tslib": "^1.9.0",
     "tslint-eslint-rules": "^5.3.1",
     "typescript-map": "0.0.7",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.7",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/language-service": "7.2.15",
+    "@angular-devkit/build-angular": "~0.803.24",
+    "@angular/cli": "8.3.24",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "8.2.14",
     "@biesbjerg/ngx-translate-extract": "^2.3.4",
     "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.116",
-    "codelyzer": "^4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^2.0.5",
@@ -62,6 +61,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "^5.9.1",
-    "typescript": "3.2.4"
+    "typescript": "3.5.3"
   }
 }
